requirements.txt
# --- Core Backend ---
Flask==3.0.3
Flask-Cors==4.0.0
Flask-PyMongo==2.3.0
pymongo==4.7.3
python-dotenv==1.0.1

# --- Authentication ---
Flask-JWT-Extended==4.6.0
bcrypt==4.2.0

# --- API & Requests ---
requests==2.32.3

# --- Production Server ---
gunicorn==22.0.0

# --- AI / ML / Image Processing ---
torch==2.3.0
transformers==4.41.2
opencv-python==4.9.0.80
pillow==10.4.0

#for LLM models
 groq